==> 01_get_good_spatials.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r}
# load libs
library(data.table)
library(sf)
```

get data

```{r}
# get good patches
good_patches <- fread("data/data_2018_good_patches.csv")

# uid
good_patches[, uid := sprintf("%i_%i_%i", id, tide_number, patch)]

# save good patches with uid
fwrite(good_patches, "data/data_2018_good_patches.csv")
# get good patch id, tide, and number
good_patches <- good_patches[, c("id", "tide_number", "patch")]
# assign marker
good_patches[, keep := T]
```

get spatial data

```{r}
# get spatials
patches <- st_read("data/patches_2018.gpkg")
```

better filter

```{r}
# get patch basic
patches <- dplyr::select(
  patches,
  id, patch, tide_number,
  time_mean, time_start, time_end
)

# merge with good patches
patches$uid <- sprintf("%i_%i_%i", patches$id, patches$tide_number, patches$patch)

# drop nas
patches <- patches[patches$uid %in% good_patches$uid, ]

# get distinct
patches <- patches[!duplicated(patches$uid), ]

# write basic
st_write(patches,
  "data/data_patches_good_2018_basic.gpkg",
  append = F
)
```
