==> 09_pairwise_association.rmd <==
---
editor_options: 
  chunk_output_type: console
---

# Pairwise associations in relation to traits

## Read in patch data and overlaps

```{r}
# to handle data
library(readr)
library(scales)
library(tidyr); library(tibble)
library(magrittr)
library(dplyr)
library(purrr)
library(stringr)
library(forcats)

# to work with networks
library(igraph)

# to plot
library(ggplot2)
library(scico)

# functions
ci <- function(x){qnorm(0.975)*sd(x, na.rm = T)/sqrt((length(x)))}
```

```{r}
# read in patch data
data_patches <- read_csv("data/data2018/data_2018_patch_summary_has_patches.csv") %>% 
  mutate(uid = as.character(1:nrow(.)))
# read in overlap data
data <- read_csv("data/data2018/data_spatio_temporal_overlap_2018.csv")
```

### Read in trait data

```{r}
# make nodes data -- this the individual identities
# add individual data to patch data
data_id <- readxl::read_excel("data/data2018/Biometrics_2018-2019.xlsx") %>% 
  filter(str_detect(`TAG NR`, "[a-zA-Z]+", negate = TRUE))

# a function for gizzard mass
get_gizzard_mass <- function(x, y) {-1.09 + (3.78*(x*y))}

# add gizzard mass
data_id <- mutate(data_id,
                  gizzard_mass = get_gizzard_mass(SH1, SW1))

# rename columns and drop ids without mass and gizzard mass
data_id <- data_id %>% 
  select(id = `TAG NR`, 
         wing = WING, mass = MASS, 
         gizzard_mass) %>% 
  distinct(id, .keep_all = TRUE) %>% 
  drop_na(gizzard_mass)

# add some exploration scores and tag info
data_behav <- read_csv("data/data2018/2018-19-all_exploration_scores.csv") %>% 
  filter(Exp == "F01")
data_tag <- read_csv("data/data2018/tag_info.csv") %>% 
  mutate(id = as.character(Toa_Tag))

# join all scores
data_id <- left_join(data_id, data_tag,
                     by = c("id")) %>% 
  left_join(data_behav, by = "FB")

# remove ids with no exploration
data_id <- mutate(data_id,
                  behav = Mean) %>% 
  # drop_na(behav) %>% 
  select(id, mass, gizzard_mass, behav)
```

### What is the empirical distribution of differences in gizzard mass?

### Filter out small patches

Each fix corresponds to 30s time.

```{r}
data_patches <- filter(data_patches,
                       nfixes > 3)

data <- data %>% 
  filter(patch_i_unique_id %in% data_patches$uid,
         patch_j_unique_id %in% data_patches$uid)
```

### Link patches with overlaps

```{r}
# convert to character
data <- mutate_at(data, vars(contains("patch")), as.character)
data <- left_join(data, data_patches,
                  by = c("patch_i_unique_id" = "uid")) %>% 
  left_join(data_patches,
            by = c("patch_j_unique_id" = "uid"))
```

### Count pairwise overlaps

```{r}
# first clip data into 3 sections with lims at 0, 55
data_summary <- data %>%
  mutate(tide_number = tide_number.x,
         waterlevel = waterlevel_start.x,
         tide_stage = case_when(waterlevel <= 0 ~ "low",
                                # between(waterlevel, 0, 55) ~ "medium",
                                waterlevel > 0 ~ "high",
                                T ~ NA_character_)) %>% 
  
  # count as well as add strength in terms of space and time
  group_by(id.x, id.y, tide_number, tide_stage) %>% 
  summarise(n_associations = length(spatial_overlap_area),
            tot_overlap_space = sum(spatial_overlap_area),
            tot_overlap_time = sum(temporal_overlap_seconds)) %>% 
  ungroup()
```

```{r}
# link trait values
data_summary <- mutate_at(data_summary,
                          vars(matches("id")),
                          as.character)
data_summary <- left_join(data_summary,
                          data_id,
                          by = c("id.x" = "id")) %>% 
  left_join(data_id,
            by = c("id.y" = "id"))

# get difference in traits
data_summary <- data_summary %>% 
  mutate(diff_gizzard = abs(gizzard_mass.x - gizzard_mass.y),
         diff_gizzard_round = plyr::round_any(diff_gizzard, 0.5),
         diff_behav = abs(behav.x - behav.y),
         tide_stage = forcats::as_factor(tide_stage),
         tide_stage = fct_relevel(tide_stage,
                                  "low", "high")) %>% 
  drop_na(diff_gizzard)
```

### Assign unique pair ids

```{r}
# count the number of data in each class
data_summary %>% 
  count(diff_gizzard_round)

# count the difference in gizzards per pairwise association
data_summary <- data_summary %>% 
  group_by(id.x, id.y) %>% 
  nest() %>% 
  ungroup() %>% 
  mutate(uid_pair = seq_len(nrow(.)))

# then unnest
data_summary <- unnest(data_summary,
                       col = "data")
```

### Subsample for even coverage

Small gizzard mass differences are over-represented as expected from a population with normally distributed traits. This requires subsampling based on difference bins.

```{r}
# filter for differences above 5 and subsample as many
# as supported
data_subsample <- data_summary %>% 
  filter(diff_gizzard_round <= 3) %>% 
  group_by(tide_stage, diff_gizzard_round) %>% 
  nest() %>%
  mutate(data = map(data, function(x) x[1:1480,])) # this is hardcoded

# unnest
data_subsample <- unnest(data_subsample, cols = "data")
```

## Statistical modelling of number of associations

### Fit a GLMM for number of associations

Difference in gizzard mass and behaviour score are predictors.

```{r}
# run a single canonical model
library(lmerTest)
model_n_assoc_full <- glmer(n_associations ~ diff_gizzard * tide_stage +
                         diff_behav * tide_stage +
                         (1|tide_number) +
                         (1|uid_pair),
                       data = data_subsample,
                       family = "poisson")

# run an alternative model with only behaviour
model_n_assoc_behav_only <- glmer(n_associations ~ 
                                 diff_behav * tide_stage +
                                 (1|tide_number) +
                                 (1|uid_pair),
                               data = data_subsample,
                               family = "poisson")

# an alternative model with only gizzard mass
model_n_assoc_gizzard_only <- glmer(n_associations ~ 
                                 diff_gizzard * tide_stage +
                                 (1|tide_number) +
                                 (1|uid_pair),
                               data = data_subsample,
                               family = "poisson")
```

### AIC comparison

Compare the AIC weights after looking at the model summaries.

```{r}
summary(model_n_assoc_full)

summary(model_n_assoc_behav_only)

summary(model_n_assoc_gizzard_only)
```

It seems that differences in behaviour as well as gizzard mass are each sufficient sufficient to explain the observed number of overlaps. Examine the AIC scores.

```{r}
# get aic difference
map_dbl(list(behaviour_only = model_n_assoc_behav_only, 
         gizzard_only = model_n_assoc_gizzard_only, 
         gizzard_and_behav = model_n_assoc_full),
    AIC)
```

With similar AIC scores and R<sup>2</sup> values, we select the reduced model by the parsimony principle, and write the model output to file.

```{r}
model_output = capture.output(map(list(behaviour_only = model_n_assoc_behav_only, 
                                       gizzard_only = model_n_assoc_gizzard_only, 
                                       gizzard_and_behav = model_n_assoc_full),
                                  summary))

# save it to file
write_lines(x = model_output, path = "results/model_n_assoc.txt")
```

Read in the model and print summary.

```{r}
cat(read_lines("results/model_n_assoc.txt"), sep = "\n")
```


## Statistical modelling of spatial overlap

### GLMM for spatial overlap

```{r}
# run a model with overlap as the response
library(lmerTest)
model_sp_overlap_full <- lmer(tot_overlap_space ~ diff_gizzard * tide_stage +
                         diff_behav * tide_stage +
                         (1|tide_number) +
                         (1|uid_pair),
                       data = data_subsample)

# run an alternative model with only behaviour
model_sp_overlap_behav_only <- lmer(tot_overlap_space ~ 
                                 diff_behav * tide_stage +
                                 (1|tide_number) +
                                 (1|uid_pair),
                               data = data_subsample)

# an alternative model with only gizzard mass
model_sp_overlap_gizzard_only <- lmer(tot_overlap_space ~ 
                                 diff_gizzard * tide_stage +
                                 (1|tide_number) +
                                 (1|uid_pair),
                               data = data_subsample)
```

### AIC comparison

Compare the AIC weights after looking at the model summaries.

```{r}
summary(model_sp_overlap_full)

summary(model_sp_overlap_behav_only)

summary(model_sp_overlap_gizzard_only)
```

It seems that differences in behaviour are sufficient in this case to explain the observed extent of overlaps. Examine the AIC scores.

```{r}
# get aic difference
map_dbl(list(behaviour_only = model_sp_overlap_behav_only, 
         gizzard_only = model_sp_overlap_gizzard_only, 
         gizzard_and_behav = model_sp_overlap_full),
    AIC)
```

```{r}
model_output = capture.output(map(list(behaviour_only = model_sp_overlap_behav_only, 
                                       gizzard_only = model_sp_overlap_gizzard_only, 
                                       gizzard_and_behav = model_sp_overlap_full),
                                  summary))

# save it to file
write_lines(x = model_output, path = "results/model_sp_overlap.txt")
```

Read in the model summary and write.

```{r}
cat(read_lines("results/model_sp_overlap.txt"), sep = "\n")
```


## Statistical modelling of temporal overlap

### GLMM for temporal overlap

```{r}
# run a model with overlap as the response
library(lmerTest)
model_time_overlap_full <- lmer(tot_overlap_time ~ diff_gizzard * tide_stage +
                         diff_behav * tide_stage +
                         (1|tide_number) +
                         (1|uid_pair),
                       data = data_subsample)

# run an alternative model with only behaviour
model_time_overlap_behav_only <- lmer(tot_overlap_time ~ 
                         diff_behav * tide_stage +
                         (1|tide_number) +
                         (1|uid_pair),
                       data = data_subsample)

# only gizzard
model_time_overlap_gizzard_only <- lmer(tot_overlap_time ~ 
                                          diff_gizzard * tide_stage +
                                          (1|tide_number) +
                                          (1|uid_pair),
                                        data = data_subsample)
```

### AIC and R<sup>2</sup> comparison

Compare the AIC weights after looking at the model summaries.

```{r}
summary(model_time_overlap_full)

summary(model_time_overlap_reduced)
```

Temporal overlap cannot be explained by the difference in gizzard mass or the difference in behaviour. This is an interesting pattern when taken together with the positive effect of behavioural difference on spatial overlap.

```{r}
# get aic difference
AIC(model_time_overlap_full) - AIC(model_time_overlap_reduced)

# get R2
MuMIn::r.squaredGLMM(model_time_overlap_reduced)
```

## Save data to file for network analyses

```{r}
write_csv(data_summary, path = "data/data_pairwise_comparisons.csv")
```
